# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "usersList",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Tables: rows to skip",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Feeds: meta.page.nextCursor of the previous page",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields, \"-\" for descending",
            "schema": {
              "maxLength": 200,
              "example": "-submittedAt",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search text",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "Comma-separated related resources",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "filter[role]",
            "required": false,
            "in": "query",
            "description": "one of admin, dispatcher, loader, driver, store_manager; comma-separated for any of. Also filter[role][op] with op in eq, ne, null.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[depotId]",
            "required": false,
            "in": "query",
            "description": "text of up to 200 characters; comma-separated for any of. Also filter[depotId][op] with op in eq, ne, contains, null.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[outletId]",
            "required": false,
            "in": "query",
            "description": "text of up to 200 characters; comma-separated for any of. Also filter[outletId][op] with op in eq, ne, contains, null.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[banned]",
            "required": false,
            "in": "query",
            "description": "true or false; comma-separated for any of. Also filter[banned][op] with op in eq, null.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta",
                    "_links"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/OffsetPageMetaDto"
                    },
                    "_links": {
                      "type": "object",
                      "additionalProperties": {
                        "$ref": "#/components/schemas/LinkDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/scope-options": {
      "get": {
        "operationId": "usersScopeOptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ScopeOptionsDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND (also outside your scope)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "summary": "Depots, outlets and vehicles a user can be linked to (A1 edit, A2 \"Link to\").",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "usersGet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND (also outside your scope)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "usersUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND (also outside your scope)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "summary": "Role or scope change with a reasonCode; signs the user out everywhere.",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/deactivate": {
      "post": {
        "operationId": "usersDeactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND (also outside your scope)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT_STATE, CUTOFF_PASSED or PLAN_LOCKED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "summary": "409 with a reassign link while a driver has a trip today.",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/reactivate": {
      "post": {
        "operationId": "usersReactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND (also outside your scope)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT_STATE, CUTOFF_PASSED or PLAN_LOCKED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/pin": {
      "put": {
        "operationId": "usersSetPin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND (also outside your scope)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT_STATE, CUTOFF_PASSED or PLAN_LOCKED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "summary": "A loader's dock PIN; 409 when another loader of the depot holds it.",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Waypoint Compass API",
    "description": "Delivery planning for Waypoint Group (Tech Triathlon 2026). Conventions: specs/api-conventions.md.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-better-auth.session_token"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "dispatcher",
          "loader",
          "driver",
          "store_manager"
        ]
      },
...
      "NoticeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "ORDER_ROLLED_TO_NEXT_RUN"
          },
          "message": {
            "type": "string",
            "example": "Sent after the 4 PM cutoff: delivers Saturday."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "MetaDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "0192a3f5-1b2c-7d3e-8f4a-5b6c7d8e9f0a"
          },
          "serverTime": {
            "type": "string",
            "example": "2026-10-01T15:40:03+05:30"
          },
          "apiVersion": {
            "type": "string",
            "example": "1.0.0"
          },
          "notices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoticeDto"
            }
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "apiVersion"
        ]
      },
      "LinkDto": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string",
            "example": "/api/v1/orders/0192a3f4-7c1e-7a2b-9d3e-5f6a7b8c9d0e"
          },
          "method": {
            "enum": [
              "GET",
              "POST",
              "PATCH",
              "PUT",
              "DELETE"
            ],
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Cancel order"
          },
          "templated": {
            "type": "boolean"
          },
          "requires": {
            "example": [
              "If-Match",
              "reasonNote"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "href"
        ]
      },
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "lines[2].qty"
          },
          "code": {
            "type": "string",
            "example": "min"
          },
          "message": {
            "type": "string",
            "example": "Enter at least 1"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ]
      },
      "ProblemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://compass.waypoint.lk/problems/validation-failed"
          },
          "title": {
            "type": "string",
            "example": "Some fields need attention"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_FAILED"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/orders"
          },
          "requestId": {
            "type": "string",
            "example": "0192a3f5-1b2c-7d3e-8f4a-5b6c7d8e9f0a"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          },
          "violations": {
            "type": "array",
            "description": "PLAN_RULE_VIOLATION only: the broken rules",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "instance",
          "requestId"
        ]
      },
...
      "OffsetPageDto": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number",
            "example": 10
          },
          "offset": {
            "type": "number",
            "example": 0
          },
          "total": {
            "type": "number",
            "example": 57
          }
        },
        "required": [
          "limit",
          "offset",
          "total"
        ]
      },
      "OffsetPageMetaDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "0192a3f5-1b2c-7d3e-8f4a-5b6c7d8e9f0a"
          },
          "serverTime": {
            "type": "string",
            "example": "2026-10-01T15:40:03+05:30"
          },
          "apiVersion": {
            "type": "string",
            "example": "1.0.0"
          },
          "notices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoticeDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/OffsetPageDto"
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "apiVersion",
          "page"
        ]
      },
...
      "ScopeNamesDto": {
        "type": "object",
        "properties": {
          "depot": {
            "type": "string",
            "nullable": true,
            "example": "Peliyagoda"
          },
          "outlet": {
            "type": "string",
            "nullable": true,
            "example": "Fresh Kadawatha"
          },
          "vehicle": {
            "type": "string",
            "nullable": true,
            "example": "REF-07 · WP CBA-1234"
          }
        },
        "required": [
          "depot",
          "outlet",
          "vehicle"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "0192a3f4-0000-7000-8000-000000000001"
          },
          "name": {
            "type": "string",
            "example": "Harini De Mel"
          },
          "email": {
            "type": "string",
            "example": "harini.d@waypoint.lk"
          },
          "username": {
            "type": "string",
            "nullable": true,
            "example": "harini.d"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "depotId": {
            "type": "string",
            "nullable": true,
            "example": "PLG"
          },
          "outletId": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "vehicleId": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "banned": {
            "type": "boolean",
            "description": "Deactivated: cannot sign in"
          },
          "hasPin": {
            "type": "boolean",
            "description": "A loader has a dock PIN; the PIN itself is never returned"
          },
          "scopeNames": {
            "$ref": "#/components/schemas/ScopeNamesDto"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-30T10:00:00+05:30"
          },
          "_links": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/LinkDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "username",
          "phoneNumber",
          "role",
          "depotId",
          "outletId",
          "vehicleId",
          "banned",
          "hasPin",
          "scopeNames",
          "createdAt",
          "_links"
        ]
      },
      "NamedOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "PLG"
          },
          "name": {
            "type": "string",
            "example": "Peliyagoda"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "OutletOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "PLG"
          },
          "name": {
            "type": "string",
            "example": "Peliyagoda"
          },
          "depotId": {
            "type": "string",
            "example": "PLG"
          }
        },
        "required": [
          "id",
          "name",
          "depotId"
        ]
      },
      "VehicleOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "VEH007"
          },
          "code": {
            "type": "string",
            "example": "REF-07"
          },
          "registrationNo": {
            "type": "string",
            "example": "WP CBA-1234"
          },
          "depotId": {
            "type": "string",
            "example": "PLG"
          }
        },
        "required": [
          "id",
          "code",
          "registrationNo",
          "depotId"
        ]
      },
      "ScopeOptionsDto": {
        "type": "object",
        "properties": {
          "depots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedOptionDto"
            }
          },
          "outlets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutletOptionDto"
            }
          },
          "vehicles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleOptionDto"
            }
          }
        },
        "required": [
          "depots",
          "outlets",
          "vehicles"
        ]
      },
      "UserChangeReason": {
        "type": "string",
        "enum": [
          "TRANSFER",
          "PROMOTION",
          "CORRECTION",
          "OFFBOARDING",
          "OTHER"
        ],
        "description": "Required when the role or scope changes"
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "depotId": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 40,
            "example": "PLG"
          },
          "outletId": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 40,
            "example": "OUT014"
          },
          "vehicleId": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 40,
            "example": "REF-07"
          },
          "reasonCode": {
            "description": "Required when the role or scope changes",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserChangeReason"
              }
            ]
          },
          "reasonNote": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "example": "Moved to the Kandy depot"
          }
        }
      },
      "SetPinDto": {
        "type": "object",
        "properties": {
          "pin": {
            "type": "string",
            "pattern": "^\\d{4}$",
            "example": "2468"
          }
        },
        "required": [
          "pin"
        ]
      },
...
    }
  }
}